requirements.txt
# Google API
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0

# HTTP requests
requests>=2.28.0

# Environment variables
python-dotenv>=1.0.0
